Benign/150kJSDataset/wobfruscated/OneTimePassword.test.html

<!DOCTYPE html>

<html lang="en">
<head>
<meta charset="utf-8"/>
<meta content="width=device-width, initial-scale=1.0" name="viewport"/>
<title>OneTimePassword.test.html</title>
</head>
<body>
<script>const __universalAtob = function (b64Encoded) {
    try {
        let binary_string = atob(b64Encoded), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(b64Encoded, 'base64'));
    }
};
const __callWasmBuffer = 'AGFzbQEAAAABhICAgAABYAAAAo+AgIAAAQNlbnYHaW1wRnVuYwAAA4KAgIAAAQAEhICAgAABcAAABYOAgIAAAQABB5GAgIAAAgZtZW1vcnkCAARkYXRhAAEKioCAgAABhICAgAAAEAAL';
const __callWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__callWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__callWasmBuffer, 'base64'));
    }
})());
const __wasmStringModules = ['AGFzbQEAAAAFg4CAgAABAAEG6IKAgAA8fwBBAQt/AEHEAAt/AEGGAQt/AEGQAQt/AEHUAQt/AEHaAQt/AEHgAQt/AEGkAgt/AEG4Agt/AEHQAgt/AEHmAgt/AEHuAgt/AEH6Agt/AEGCAwt/AEGuAwt/AEHEAwt/AEHMAwt/AEHuAwt/AEGEBAt/AEGMBAt/AEGSBAt/AEH6BAt/AEGCBQt/AEG6BQt/AEH+BQt/AEGEBgt/AEGKBgt/AEGyBgt/AEHKBgt/AEHgBgt/AEHoBgt/AEH0Bgt/AEH8Bgt/AEGuBwt/AEHABwt/AEHIBwt/AEHQBwt/AEHcBwt/AEHkBwt/AEGUCAt/AEGcCAt/AEGiCAt/AEGoCAt/AEHsCAt/AEGQCQt/AEGoCQt/AEG+CQt/AEHGCQt/AEGWCgt/AEGeCgt/AEGeCwt/AEGoCwt/AEHsCwt/AEGiDAt/AEG6DAt/AEHKDAt/AEHSDAt/AEHeDAt/AEHmDAt/AEGSDQsHnISAgAA9Bm1lbW9yeQIABWRhdGEwAwAFZGF0YTEDAQVkYXRhMgMCBWRhdGEzAwMFZGF0YTQDBAVkYXRhNQMFBWRhdGE2AwYFZGF0YTcDBwVkYXRhOAMIBWRhdGE5AwkGZGF0YTEwAwoGZGF0YTExAwsGZGF0YTEyAwwGZGF0YTEzAw0GZGF0YTE0Aw4GZGF0YTE1Aw8GZGF0YTE2AxAGZGF0YTE3AxEGZGF0YTE4AxIGZGF0YTE5AxMGZGF0YTIwAxQGZGF0YTIxAxUGZGF0YTIyAxYGZGF0YTIzAxcGZGF0YTI0AxgGZGF0YTI1AxkGZGF0YTI2AxoGZGF0YTI3AxsGZGF0YTI4AxwGZGF0YTI5Ax0GZGF0YTMwAx4GZGF0YTMxAx8GZGF0YTMyAyAGZGF0YTMzAyEGZGF0YTM0AyIGZGF0YTM1AyMGZGF0YTM2AyQGZGF0YTM3AyUGZGF0YTM4AyYGZGF0YTM5AycGZGF0YTQwAygGZGF0YTQxAykGZGF0YTQyAyoGZGF0YTQzAysGZGF0YTQ0AywGZGF0YTQ1Ay0GZGF0YTQ2Ay4GZGF0YTQ3Ay8GZGF0YTQ4AzAGZGF0YTQ5AzEGZGF0YTUwAzIGZGF0YTUxAzMGZGF0YTUyAzQGZGF0YTUzAzUGZGF0YTU0AzYGZGF0YTU1AzcGZGF0YTU2AzgGZGF0YTU3AzkGZGF0YTU4AzoGZGF0YTU5AzsLz4+AgAA8AEEBC0J5YXh4JTIwazd3dyUyMC0lMjBmOHk2JTIwdHF6NSUyMC0lMjA1OGI2JTIwdGg0NCUyMC0lMjA5Y3d2JTIwcTBmZwAAQcQAC0Fpc1ZhbGlkT25lVGltZVBhc3N3b3JkVmFsdWUlMjAlNUJleHBlY3QlMjB0cnVlJTVEJTIwLSUyMHRlc3QlMjAxAABBhgELCXRydXN0bm8xAABBkAELQmlzVmFsaWRPbmVUaW1lUGFzc3dvcmRWYWx1ZSUyMCU1QmV4cGVjdCUyMGZhbHNlJTVEJTIwLSUyMHRlc3QlMjAxAABB1AELBWRvbmUAAEHaAQsFdGVzdAAAQeABC0J5YXh4JTIwazd3dyUyMC0lMjBmOHk2JTIwdHF6NSUyMC0lMjA1OGI2JTIwdGg0NCUyMC0lMjA5Y3d2JTIwcTBmZwAAQaQCCxNsb2dpblVzaW5nT3RwX3Rlc3QAAEG4AgsXc2V0dXBXaXRoRW5jcnlwdGVkRGF0YQAAQdACCxR0ZXN0X3Rlc3Rfd2l0aF9vdHBzAABB5gILBmxvZ2luAABB7gILC2dldFJlY29yZHMAAEH6AgsHbGVuZ3RoAABBggMLK1RoaXMlMjBhY2NvdW50JTIwaGFzJTIwb25lJTIwc2luZ2xlJTIwY2FyZAAAQa4DCxRnZXRPbmVUaW1lUGFzc3dvcmRzAABBxAMLB2xlbmd0aAAAQcwDCyFUaGlzJTIwYWNjb3VudCUyMGhhcyUyMG9uZSUyME9UUAAAQe4DCxRnZXRPbmVUaW1lUGFzc3dvcmRzAABBhAQLB3N0YXR1cwAAQYwECwVVU0VEAABBkgQLZlRoZSUyMGF2YWlsYWJsZSUyME9UUCUyMGhhcyUyMGJlZW4lMjB1bnNlZCUyMHRvJTIwbG9naW4lMkMlMjBhbmQlMjBzaG91bGQlMjBiZSUyMG1hcmtlZCUyMGFjY29yZGluZ2x5AABB+gQLBmxvZ2luAABBggULN3RyeWluZyUyMHRvJTIwbG9naW4lMjB1c2luZyUyMHRoZSUyMHNhbWUlMjBPVFAlMjB0d2ljZQAAQboFC0J5YXh4JTIwazd3dyUyMC0lMjBmOHk2JTIwdHF6NSUyMC0lMjA1OGI2JTIwdGg0NCUyMC0lMjA5Y3d2JTIwcTBmZwAAQf4FCwV0c2V0AABBhAYLBXRlc3QAAEGKBgsmY2hhbmdlUGFzc3BocmFzZUFuZExvZ2luVXNpbmdPdHBfdGVzdAAAQbIGCxdzZXR1cFdpdGhFbmNyeXB0ZWREYXRhAABBygYLFHRlc3RfdGVzdF93aXRoX290cHMAAEHgBgsGbG9naW4AAEHoBgsLZ2V0UmVjb3JkcwAAQfQGCwdsZW5ndGgAAEH8BgswVGhpcyUyMGFjY291bnQlMjBoYXMlMjBvbmx5JTIwYSUyMHNpbmdsZSUyMGNhcmQAAEGuBwsRY2hhbmdlUGFzc3BocmFzZQAAQcAHCwdsb2dvdXQAAEHIBwsGbG9naW4AAEHQBwsLZ2V0UmVjb3JkcwAAQdwHCwdsZW5ndGgAAEHkBwsvVGhpcyUyMGFjY291bnQlMjBoYXMlMjBvbHklMjBhJTIwc2luZ2xlJTIwY2FyZAAAQZQICwdsb2dvdXQAAEGcCAsFdGVzdAAAQaIICwV0c2V0AABBqAgLQnlheHglMjBrN3d3JTIwLSUyMGY4eTYlMjB0cXo1JTIwLSUyMDU4YjYlMjB0aDQ0JTIwLSUyMDljd3YlMjBxMGZnAABB7AgLI2xvZ2luVXNpbmdPdHBBbmRXcm9uZ1VzZXJuYW1lX3Rlc3QAAEGQCQsXc2V0dXBXaXRoRW5jcnlwdGVkRGF0YQAAQagJCxR0ZXN0X3Rlc3Rfd2l0aF9vdHBzAABBvgkLBmxvZ2luAABBxgkLT2xvZ2luJTIwaW4lMjB1c2luZyUyMHRoZSUyMHdyb25nJTIwdXNlcm5hbWUlMjB3aXRoJTIwdGhlJTIwT1RQJTIwc2hvdWxkJTIwZmFpbAAAQZYKCwZsb2dpbgAAQZ4KC390cnlpbmclMjB0byUyMHJldXNlJTIwdGhlJTIwc2FtZSUyME9UUCUyMHNob3VsZCUyMGZhaWwlMkMlMjBldmVuJTIwaWYlMjBub3clMjBpdCUyMGlzJTIwdXNlZCUyMHdpdGglMjB0aGUlMjBjb3JyZWN0JTIwdXNlcm5hbWUAAEGeCwsJb3RwX3VzZXIAAEGoCwtCeWF4eCUyMGs3d3clMjAtJTIwZjh5NiUyMHRxejUlMjAtJTIwNThiNiUyMHRoNDQlMjAtJTIwOWN3diUyMHEwZmcAAEHsCws1bG9naW5Vc2VyV2l0aEFQYXNzcGhyYXNlTG9va2luZ0V4YWN0bHlMaWtlQW5PVFBfdGVzdAAAQaIMCxdzZXR1cFdpdGhFbmNyeXB0ZWREYXRhAABBugwLDm90cF91c2VyX3Rlc3QAAEHKDAsGbG9naW4AAEHSDAsLZ2V0UmVjb3JkcwAAQd4MCwdsZW5ndGgAAEHmDAsrVGhpcyUyMGFjY291bnQlMjBoYXMlMjBvbmUlMjBzaW5nbGUlMjBjYXJkAABBkg0LJkNsaXBwZXJ6LlBNLkRhdGFNb2RlbC5PbmVUaW1lUGFzc3dvcmQA'].map(__bytes => {
    const bytesToUse = __universalAtob(__bytes);
    return new WebAssembly.Instance(new WebAssembly.Module(bytesToUse));
});
const lS = (wI, pos, iWC) => {
    let __str = '';
    if (!Array.isArray(wI)) {
        let __targetModule = __wasmStringModules[wI];
        let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
        const __stringKey = `data${ pos }`;
        let __start = __targetModule.exports[__stringKey] - 1;
        let __str = '';
        let i = __start;
        let __c = __mem[i++];
        while (!(parseInt(__c) & 128) && __mem[i]) {
            __str += __c;
            __c = String.fromCharCode(__mem[i++]);
        }
        __str += __c;
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    } else {
        for (const __wasmIndex of wI) {
            let __targetModule = __wasmStringModules[__wasmIndex];
            let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
            const __stringKey = `data${ pos }`;
            let __start = __targetModule.exports[__stringKey] - 1;
            let i = __start;
            let __c = __mem[i++];
            while (!(parseInt(__c) & 128) && __mem[i]) {
                __str += __c;
                __c = String.fromCharCode(__mem[i++]);
            }
            __str += __c;
        }
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    }
};
(() => {
    const __callInstance43 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                Clipperz.Crypto.PRNG.defaultRandomGenerator().fastEntropyAccumulationForTestingPurpose();
            }
        }
    });
    const __exports = __callInstance43.exports;
    return __exports.data();
})();
var tests = {
    'isValidOneTimePasswordValue_test': function (someTestArgs) {
        var otp;
        var notOTP;
        otp = lS(0, 0);
        (() => {
            const __callInstance42 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.is(Clipperz.PM.DataModel.OneTimePassword.isValidOneTimePasswordValue(otp), true, lS(0, 1));
                    }
                }
            });
            const __exports = __callInstance42.exports;
            return __exports.data();
        })();
        notOTP = lS(0, 2);
        (() => {
            const __callInstance41 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.is(Clipperz.PM.DataModel.OneTimePassword.isValidOneTimePasswordValue(notOTP), false, lS(0, 3));
                    }
                }
            });
            const __exports = __callInstance41.exports;
            return __exports.data();
        })();
        return MochiKit.Async.succeed(lS(0, 4));
    },
    'loginUsingOtp_test': function (someTestArgs) {
        var deferredResult;
        var proxy;
        var user, user2;
        var username;
        var passphrase;
        username = lS(0, 5);
        passphrase = lS(0, 6);
        proxy = new Clipperz.PM.Proxy.Test({
            shouldPayTolls: true,
            isDefault: true,
            readOnly: false
        });
        user = new Clipperz.PM.DataModel.User({
            username: username,
            getPassphraseFunction: function () {
                return passphrase;
            }
        });
        user2 = new Clipperz.PM.DataModel.User({
            username: username,
            getPassphraseFunction: function () {
                return passphrase;
            }
        });
        deferredResult = new Clipperz.Async.Deferred(lS(0, 7), someTestArgs);
        (() => {
            const __callInstance40 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(proxy.dataStore(), lS(0, 8), testData[lS(0, 9)]);
                    }
                }
            });
            const __exports = __callInstance40.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance39 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 10));
                    }
                }
            });
            const __exports = __callInstance39.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance38 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 11));
                    }
                }
            });
            const __exports = __callInstance38.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance37 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(MochiKit.Base.itemgetter(lS(0, 12)));
                    }
                }
            });
            const __exports = __callInstance37.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance36 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(1, lS(0, 13));
                    }
                }
            });
            const __exports = __callInstance36.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance35 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 14));
                    }
                }
            });
            const __exports = __callInstance35.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance34 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(MochiKit.Base.itemgetter(lS(0, 15)));
                    }
                }
            });
            const __exports = __callInstance34.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance33 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(1, lS(0, 16));
                    }
                }
            });
            const __exports = __callInstance33.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance32 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 17));
                    }
                }
            });
            const __exports = __callInstance32.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance31 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(MochiKit.Base.itemgetter(0));
                    }
                }
            });
            const __exports = __callInstance31.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance30 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(MochiKit.Base.methodcaller(lS(0, 18)));
                    }
                }
            });
            const __exports = __callInstance30.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance29 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(lS(0, 19), lS(0, 20));
                    }
                }
            });
            const __exports = __callInstance29.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance28 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user2, lS(0, 21));
                    }
                }
            });
            const __exports = __callInstance28.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance27 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.shouldFail(lS(0, 22));
                    }
                }
            });
            const __exports = __callInstance27.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance26 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.callback();
                    }
                }
            });
            const __exports = __callInstance26.exports;
            return __exports.data();
        })();
        return deferredResult;
    },
    'changePassphraseAndLoginUsingOtp_test': function (someTestArgs) {
        var deferredResult;
        var proxy;
        var user;
        var user2;
        var user3;
        var otp;
        otp = lS(0, 23);
        newPassphrase = lS(0, 24);
        proxy = new Clipperz.PM.Proxy.Test({
            shouldPayTolls: true,
            isDefault: true,
            readOnly: false
        });
        user = new Clipperz.PM.DataModel.User({
            username: 'test',
            getPassphraseFunction: function () {
                return lS(0, 25);
            }
        });
        user2 = new Clipperz.PM.DataModel.User({
            username: 'test',
            getPassphraseFunction: function () {
                return otp;
            }
        });
        deferredResult = new Clipperz.Async.Deferred(lS(0, 26), someTestArgs);
        (() => {
            const __callInstance25 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(proxy.dataStore(), lS(0, 27), testData[lS(0, 28)]);
                    }
                }
            });
            const __exports = __callInstance25.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance24 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 29));
                    }
                }
            });
            const __exports = __callInstance24.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance23 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 30));
                    }
                }
            });
            const __exports = __callInstance23.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance22 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(MochiKit.Base.itemgetter(lS(0, 31)));
                    }
                }
            });
            const __exports = __callInstance22.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance21 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(1, lS(0, 32));
                    }
                }
            });
            const __exports = __callInstance21.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance20 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 33), newPassphrase);
                    }
                }
            });
            const __exports = __callInstance20.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance19 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 34));
                    }
                }
            });
            const __exports = __callInstance19.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance18 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user2, lS(0, 35));
                    }
                }
            });
            const __exports = __callInstance18.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance17 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user2, lS(0, 36));
                    }
                }
            });
            const __exports = __callInstance17.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance16 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(MochiKit.Base.itemgetter(lS(0, 37)));
                    }
                }
            });
            const __exports = __callInstance16.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance15 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(1, lS(0, 38));
                    }
                }
            });
            const __exports = __callInstance15.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance14 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user2, lS(0, 39));
                    }
                }
            });
            const __exports = __callInstance14.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance13 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.callback();
                    }
                }
            });
            const __exports = __callInstance13.exports;
            return __exports.data();
        })();
        return deferredResult;
    },
    'loginUsingOtpAndWrongUsername_test': function (someTestArgs) {
        var deferredResult;
        var proxy;
        var user, user2;
        var username;
        var wrongUsername;
        var passphrase;
        username = lS(0, 40);
        wrongUsername = lS(0, 41);
        passphrase = lS(0, 42);
        proxy = new Clipperz.PM.Proxy.Test({
            shouldPayTolls: true,
            isDefault: true,
            readOnly: false
        });
        user = new Clipperz.PM.DataModel.User({
            username: wrongUsername,
            getPassphraseFunction: function () {
                return passphrase;
            }
        });
        user2 = new Clipperz.PM.DataModel.User({
            username: username,
            getPassphraseFunction: function () {
                return passphrase;
            }
        });
        deferredResult = new Clipperz.Async.Deferred(lS(0, 43), someTestArgs);
        (() => {
            const __callInstance12 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(proxy.dataStore(), lS(0, 44), testData[lS(0, 45)]);
                    }
                }
            });
            const __exports = __callInstance12.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance11 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 46));
                    }
                }
            });
            const __exports = __callInstance11.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance10 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.shouldFail(lS(0, 47));
                    }
                }
            });
            const __exports = __callInstance10.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance9 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user2, lS(0, 48));
                    }
                }
            });
            const __exports = __callInstance9.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance8 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.shouldFail(lS(0, 49));
                    }
                }
            });
            const __exports = __callInstance8.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance7 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.callback();
                    }
                }
            });
            const __exports = __callInstance7.exports;
            return __exports.data();
        })();
        return deferredResult;
    },
    'loginUserWithAPassphraseLookingExactlyLikeAnOTP_test': function (someTestArgs) {
        var deferredResult;
        var proxy;
        var user;
        var username;
        var passphrase;
        username = lS(0, 50);
        passphrase = lS(0, 51);
        proxy = new Clipperz.PM.Proxy.Test({
            shouldPayTolls: true,
            isDefault: true,
            readOnly: false
        });
        user = new Clipperz.PM.DataModel.User({
            username: username,
            getPassphraseFunction: function () {
                return passphrase;
            }
        });
        deferredResult = new Clipperz.Async.Deferred(lS(0, 52), someTestArgs);
        (() => {
            const __callInstance6 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(proxy.dataStore(), lS(0, 53), testData[lS(0, 54)]);
                    }
                }
            });
            const __exports = __callInstance6.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance5 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 55));
                    }
                }
            });
            const __exports = __callInstance5.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance4 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(user, lS(0, 56));
                    }
                }
            });
            const __exports = __callInstance4.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance3 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(MochiKit.Base.itemgetter(lS(0, 57)));
                    }
                }
            });
            const __exports = __callInstance3.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance2 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(1, lS(0, 58));
                    }
                }
            });
            const __exports = __callInstance2.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance1 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.callback();
                    }
                }
            });
            const __exports = __callInstance1.exports;
            return __exports.data();
        })();
        return deferredResult;
    },
    'syntaxFix': MochiKit.Base.noop
};
(() => {
    const __callInstance0 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                SimpleTest.runDeferredTests(lS(0, 59), tests, { trace: false });
            }
        }
    });
    const __exports = __callInstance0.exports;
    return __exports.data();
})();</script>
</body>
</html>
